+++ src/index.html
@@ -8,13 +8,14 @@
     <link rel="stylesheet" href="styles.css">
   </head>
   <body>
-    <script type="module" src="main.js"></script>
-    <script type="module" src="lib/login.js"></script>
-    <script type="module" src="./firebase.js"></script>
+    
     <div id="root"></div>   
     <!-- TODO ESTO ES LA INSTALACION DE FIREBASE-->
     <script src="https://www.gstatic.com/firebasejs/8.2.7/firebase-app.js"></script>
     <script src="https://www.gstatic.com/firebasejs/8.2.7/firebase-auth.js"></script>
     <script src="https://www.gstatic.com/firebasejs/8.2.7/firebase-firestore.js"></script>
+    <script type="module" src="main.js"></script>
+    <script type="module" src="lib/login.js"></script>
+    <script type="module" src="./firebase.js"></script>
   </body>
 </html>
