+++ src/main.js
@@ -1,6 +1,6 @@
 import { onNavigate } from './routers.js';
-import { register, loginGoogle, accessJalo, historyRef } from './firebase.js';
-//import { cardWall } from './lib/card-wall.js';
+import { register, loginGoogle, accessJalo, deleteHistory, getHistoryEdit} from './firebase.js';
+import { cardWall } from './lib/card-wall.js';
 
 //Función para mandar llamar el id que se usa para el evento para ir de home a login.
 const createNewUser = () => {
@@ -27,7 +27,7 @@ window.addEventListener('DOMContentLoaded', () => oldUser1());
 
 
 
-//login a wall
+//login to wall
 const buttonLogin = () => {
     let youLogin = document.getElementById('checkIn');
     youLogin.addEventListener('click', (e) => {
@@ -40,7 +40,7 @@ const buttonLogin = () => {
 window.addEventListener('DOMContentLoaded', () => buttonLogin());
 
 
-//Google a wall
+//Google to wall
 const buttonGoogle = () => {
     let youLoginGoogle = document.getElementById('buttonGoogle');
     youLoginGoogle.addEventListener('click', (e) => {
@@ -51,7 +51,7 @@ const buttonGoogle = () => {
 window.addEventListener('DOMContentLoaded', () => buttonGoogle());
 
 
-//Google a wall con inputs
+//Google a wall with inputs
 const buttonGoogleInput = () => {
     let youLoginGoogleInputs = document.getElementById('buttonLoginInputs');
     youLoginGoogleInputs.addEventListener('click', (e) => {
@@ -72,5 +72,46 @@ buttonHistories.addEventListener('click', (e) => {
     
     historyRef(title, description);
     console.log(title, description);
-    buttonHistories.resert();
+    buttonHistories.reset();
 });
+
+
+
+//put all the histories and delete
+let praintCards = document.querySelector('#tasks-container');
+export const setupPost = data => {
+  if (data.length) {
+    let html = '';
+    data.forEach(doc => {
+        const post = doc.data()
+        post.id = doc.id;
+        const praint = cardWall(post);
+        html += praint;
+        });
+    praintCards.innerHTML=html;
+    } else {
+    praintCards.innerHTML='<p>Login to see Posts</p>';
+    };
+
+const buttonDelete = document.querySelectorAll('.deletePublication');
+buttonDelete.forEach(history => {
+    history.addEventListener('click', (e) => {
+         deleteHistory(e.target.dataset.id);
+        })
+    })
+
+const buttonEdit = document.querySelectorAll('.editPublication');
+buttonEdit.forEach(history => {
+    history.addEventListener('click', async (e) => {
+        console.log('editando'); 
+        console.log(e.target.dataset.id);
+        const doc = await getHistoryEdit(e.target.dataset.id);
+        console.log(doc.data());
+        //editeHistory(e.target.dataset.id);
+        })
+    })
+};
+
+
+
+
